fourth proposal for UI change
