+++ packages/bruno-app/src/components/Actions/index.js
@@ -1,25 +1,51 @@
-import StyledWrapper from './StyledWrapper';
 import toast from 'react-hot-toast';
+import StyledWrapper from './StyledWrapper';
 
-import { useState, forwardRef, useRef } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
-import { showHomePage } from 'providers/ReduxStore/slices/app';
-import { IconSettings, IconPlus, IconFolders, IconDownload, IconActivity } from '@tabler/icons';
-import { showPreferences } from 'providers/ReduxStore/slices/app';
-import { openCollection, importCollection } from 'providers/ReduxStore/slices/collections/actions';
+import {
+  IconArrowsSort,
+  IconDownload,
+  IconFolders,
+  IconPlus,
+  IconSortAscendingLetters,
+  IconSortDescendingLetters
+} from '@tabler/icons';
+import { IconSearch, IconX } from '@tabler/icons';
+import Preferences from 'components/Preferences/index';
 import CreateCollection from 'components/Sidebar/CreateCollection/index';
 import ImportCollection from 'components/Sidebar/ImportCollection/index';
 import ImportCollectionLocation from 'components/Sidebar/ImportCollectionLocation/index';
-import Preferences from 'components/Preferences/index';
+import { showPreferences } from 'providers/ReduxStore/slices/app';
+import { importCollection, openCollection, sortCollections } from 'providers/ReduxStore/slices/collections/actions';
+import { useState } from 'react';
+import { useDispatch, useSelector } from 'react-redux';
 
 const Actions = () => {
   const [importedCollection, setImportedCollection] = useState(null);
   const [createCollectionModalOpen, setCreateCollectionModalOpen] = useState(false);
   const [importCollectionModalOpen, setImportCollectionModalOpen] = useState(false);
+  const [searchText, setSearchText] = useState('');
   const [importCollectionLocationModalOpen, setImportCollectionLocationModalOpen] = useState(false);
   const preferencesOpen = useSelector((state) => state.app.showPreferences);
   const dispatch = useDispatch();
   const { ipcRenderer } = window;
+
+  const { collections } = useSelector((state) => state.collections);
+  const { collectionSortOrder } = useSelector((state) => state.collections);
+  const sortCollectionOrder = () => {
+    let order;
+    switch (collectionSortOrder) {
+      case 'default':
+        order = 'alphabetical';
+        break;
+      case 'alphabetical':
+        order = 'reverseAlphabetical';
+        break;
+      case 'reverseAlphabetical':
+        order = 'default';
+        break;
+    }
+    dispatch(sortCollections({ order }));
+  };
 
   const handleImportCollection = (collection) => {
     setImportedCollection(collection);
@@ -40,10 +66,6 @@ const Actions = () => {
     );
   };
 
-  const openDevTools = () => {
-    ipcRenderer.invoke('renderer:open-devtools');
-  };
-
   return (
     <StyledWrapper className="px-2 py-1 flex flex-row">
       {createCollectionModalOpen ? <CreateCollection onClose={() => setCreateCollectionModalOpen(false)} /> : null}
@@ -60,7 +82,7 @@ const Actions = () => {
       {preferencesOpen && <Preferences onClose={() => dispatch(showPreferences(false))} />}
       <div className="group flex-grow">
         <span className="group-title">Collections</span>
-        <div className="flex flex-row justify-center">
+        <div className="flex flex-wrap justify-center">
           <div className="px-1 py-2">
             <button
               type="button"
@@ -100,34 +122,48 @@ const Actions = () => {
               Import
             </button>
           </div>
-        </div>
-      </div>
-      <div className="group flex-grow">
-        <span className="group-title">Settings</span>
-        <div className="flex flex-row justify-center">
-          <div className="px-1 py-2 ">
-            <button
-              type="button"
-              title="Preferences"
-              className="btn btn-secondary btn-xs flex flex-row"
-              onClick={(e) => {
-                dispatch(showPreferences(true));
-              }}
-            >
-              <IconSettings strokeWidth={1.5} />
+          {collections.length >= 1 && (
+            <button onClick={() => sortCollectionOrder()}>
+              {collectionSortOrder == 'default' ? (
+                <IconArrowsSort size={18} strokeWidth={1.5} />
+              ) : collectionSortOrder == 'alphabetical' ? (
+                <IconSortAscendingLetters size={18} strokeWidth={1.5} />
+              ) : (
+                <IconSortDescendingLetters size={18} strokeWidth={1.5} />
+              )}
             </button>
-          </div>
-          <div className="py-2 ">
-            <button
-              type="button"
-              className="btn btn-xs flex flex-row"
-              title="Devtools"
-              onClick={(e) => {
-                openDevTools();
-              }}
-            >
-              <IconActivity strokeWidth={1.5} />
-            </button>
+          )}
+          <div className=" relative collection-filter py-2 px-1 flex-grow">
+            <div className="absolute inset-y-0 left-0 pl-4 flex items-center pointer-events-none">
+              <span className="text-gray-500 sm:text-sm">
+                <IconSearch size={16} strokeWidth={1.5} />
+              </span>
+            </div>
+            <input
+              type="text"
+              name="search"
+              id="search"
+              autoComplete="off"
+              autoCorrect="off"
+              autoCapitalize="off"
+              spellCheck="false"
+              className="block w-full pl-7 py-1 sm:text-sm"
+              placeholder="search"
+              value={searchText}
+              onChange={(e) => setSearchText(e.target.value.toLowerCase())}
+            />
+            {searchText !== '' && (
+              <div className="absolute inset-y-0 right-0 pr-4 flex items-center">
+                <span
+                  className="close-icon"
+                  onClick={() => {
+                    setSearchText('');
+                  }}
+                >
+                  <IconX size={16} strokeWidth={1.5} className="cursor-pointer" />
+                </span>
+              </div>
+            )}
           </div>
         </div>
       </div>
