+++ packages/bruno-app/src/components/Sidebar/TitleBar/StyledWrapper.js
@@ -14,6 +14,16 @@ const StyledWrapper = styled.div`
       user-select: none;
     }
   }
+  .group {
+    border: solid 2px;
+    border-color: ${(props) => props.theme.dropdown.bg};
+    border-radius: 5px;
+    margin-right: 2px;
+    display: grid;
+  }
+  .group-title {
+    text-align: center;
+  }
 `;
 
 export default StyledWrapper;
